Split file.h into more headers
Now it's a bit more clear to understand and I'll expand the new headers
with more functions
